requirements.txt
flask==2.0.1
line-bot-sdk==2.0.1
schedule==1.1.0
requests==2.26.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
werkzeug==2.0.3
pytz==2022.1
Pillow==9.2.0
pandas==1.3.5
openpyxl==3.0.9
python-dateutil==2.8.2
